refactor(analysis): own structured report data
